+++ HISTORY.rst
@@ -35,6 +35,7 @@ History
     * Improve LeadOrder model unittests (rudaporto).
     * Card #273: Added new state to Assignment: post_processing (rudaporto).
     * Card #273: Added new transitions to move to and back in_qa to post_processing and to approve from post_processing (rudaporto).
+    * Reclassify Report views to be marked as background tasks in newrelic agent (rudaporto).
 
 2.1.12 (2017-04-28)
 -------------------